% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_object_exist}
\alias{check_object_exist}
\title{检查对象是否存在}
\usage{
check_object_exist(config = cfg_edge, object_name = "t_icitem")
}
\arguments{
\item{config}{连接}

\item{object_name}{对象名}
}
\value{
返回值
}
\description{
检查对象是否存在
}
\examples{
check_object_exist()
}
