% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_clear_input}
\alias{cloud_clear_input}
\title{云端清除缓存表}
\usage{
cloud_clear_input(
  config = config_cloud,
  input_tableName = "rds_icitem_view3_input"
)
}
\arguments{
\item{config}{配置表}

\item{input_tableName}{输入表名}
}
\value{
返回值
}
\description{
云端清除缓存表
}
\examples{
cloud_clear_input()
}
