% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_createTable_data}
\alias{cloud_createTable_data}
\title{在云边创建表结果}
\usage{
cloud_createTable_data(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  src_edge_viewName = "t_icitem_view3",
  target_cloud_tableName = "rds_icitem_view3",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config_edge}{边配置文件}

\item{config_cloud}{云配置文件}

\item{src_edge_viewName}{边视图文件}

\item{target_cloud_tableName}{云表文件}

\item{field_selector}{字段选择器}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
在云边创建表结果
}
\examples{
cloud_createTable_data()
}
