% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_createTable_input}
\alias{cloud_createTable_input}
\title{创建日志日志输入的文件}
\usage{
cloud_createTable_input(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  src_edge_diff_tableName = "t_icitem_view3_diff",
  target_cloud_input_tableName = "rds_icitem_view3_input",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config_edge}{边配置文件}

\item{config_cloud}{云配置文件}

\item{src_edge_diff_tableName}{边差异表包含Action}

\item{target_cloud_input_tableName}{云配置表}

\item{field_selector}{字段选择器}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
创建日志日志输入的文件
}
\examples{
cloud_createTable_input()
}
