% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_createTable_log}
\alias{cloud_createTable_log}
\title{创建云边日志文件}
\usage{
cloud_createTable_log(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  src_edge_log_tableName = "t_icitem_view3_log",
  target_cloud_log_tableName = "rds_icitem_view3_log",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config_edge}{边配置文件}

\item{config_cloud}{云配置文件}

\item{src_edge_log_tableName}{边日志表}

\item{target_cloud_log_tableName}{云日志表}

\item{field_selector}{字段选择器}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
创建云边日志文件
}
\examples{
cloud_createTable_log()
}
