% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_log_setDone}
\alias{cloud_log_setDone}
\title{设置云端日志状态}
\usage{
cloud_log_setDone(
  config = config_cloud,
  log_tableName = "rds_icitem_view3_log"
)
}
\arguments{
\item{config}{配置}

\item{log_tableName}{日志表名}
}
\value{
返回值
}
\description{
设置云端日志状态
}
\examples{
cloud_log_setDone()
}
