% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_sync_add}
\alias{cloud_sync_add}
\title{同步云端新增操作}
\usage{
cloud_sync_add(
  config = config_cloud,
  data_tableName = "rds_icitem_view3",
  input_tableName = "rds_icitem_view3_input",
  FDiffAction = "add"
)
}
\arguments{
\item{config}{配置}

\item{data_tableName}{数据表}

\item{input_tableName}{输入表名}

\item{FDiffAction}{操作标识}
}
\value{
返回值
}
\description{
同步云端新增操作
}
\examples{
cloud_sync_add()
}
