% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_sync_delete}
\alias{cloud_sync_delete}
\title{同步云数据删除部分}
\usage{
cloud_sync_delete(
  config = config_cloud,
  data_tableName = "rds_icitem_view3",
  input_tableName = "rds_icitem_view3_input",
  FInterID = "FItemID",
  FDiffAction = "delete"
)
}
\arguments{
\item{config}{配置}

\item{data_tableName}{数据表}

\item{input_tableName}{输入表}

\item{FInterID}{内码}

\item{FDiffAction}{操作码}
}
\value{
返回值
}
\description{
同步云数据删除部分
}
\examples{
cloud_sync_delete()
}
