% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cloud_tableInfo_init}
\alias{cloud_tableInfo_init}
\title{添加对云边数据库的初始化}
\usage{
cloud_tableInfo_init(config = cfg_cloud)
}
\arguments{
\item{config}{云数据库配置}
}
\value{
返回值
}
\description{
添加对云边数据库的初始化
}
\examples{
cloud_tableInfo_init()
}
