% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dbo_func_statField}
\alias{dbo_func_statField}
\title{返回日志文件的最大版本}
\usage{
dbo_func_statField(
  config = cfg_edge,
  table_name = "t_icitem_log",
  return_field = "FVersion",
  return_type = "int",
  stat_func = "max",
  step = 1,
  func_name = "func_icitem_max_ver"
)
}
\arguments{
\item{config}{配置}

\item{table_name}{表名}

\item{return_field}{返回值字值}

\item{return_type}{返回值类型默认为int或者float}

\item{stat_func}{统计函数默认为max}

\item{step}{步骤默认为1也可以是-1}

\item{func_name}{处理函数默认最大值}
}
\value{
返回值
}
\description{
返回日志文件的最大版本
}
\examples{
dbo_func_statField()
}
