% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_data_del}
\alias{edge_data_del}
\title{边数据删除备份表}
\usage{
edge_data_del(
  config = cfg_edge,
  src_view_name = "t_icitem_view3",
  target_table_name = "t_icitem_del",
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{src_view_name}{源数据}

\item{target_table_name}{删除备份表}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
边数据删除备份表
}
\examples{
edge_data_del()
}
