% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_data_img}
\alias{edge_data_img}
\title{增加边的镜像文件}
\usage{
edge_data_img(
  config = cfg_edge,
  src_view_name = "t_icitem_view3",
  target_table_name = "t_icitem_img",
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{src_view_name}{数据源}

\item{target_table_name}{镜像表名}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
增加边的镜像文件
}
\examples{
edge_data_img()
}
