% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_data_view}
\alias{edge_data_view}
\title{针对边数据进行定义}
\usage{
edge_data_view(
  config = cfg_edge,
  table_name = "t_icitem",
  view_name = "t_icitem_view",
  selector = c("FItemID", "FName")
)
}
\arguments{
\item{config}{边配置}

\item{table_name}{表名}

\item{selector}{选择器}

\item{view_suffix}{后缀默认为_view}
}
\value{
返回值
}
\description{
针对边数据进行定义
}
\examples{
edge_data_view()
}
