% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_diff_add}
\alias{edge_diff_add}
\title{差异新增}
\usage{
edge_diff_add(
  config = cfg_edge,
  src_view_name = "t_icitem_view3",
  img_table_name = "t_icitem_view3_img",
  target_view_name = "t_icitem_view3_diff_add",
  key_col = "FItemId"
)
}
\arguments{
\item{config}{配置}

\item{src_view_name}{数据源}

\item{img_table_name}{镜像数据}

\item{target_view_name}{目标视图}

\item{key_col}{主键}
}
\value{
返回键
}
\description{
差异新增
}
\examples{
edge_diff_add()
}
