% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_diff_del}
\alias{edge_diff_del}
\title{测试删除数据}
\usage{
edge_diff_del(
  config = cfg_edge,
  src_view_name = "t_icitem_view3",
  img_table_name = "t_icitem_view3_img",
  target_view_name = "t_icitem_view3_diff_del",
  key_col = "FItemId"
)
}
\arguments{
\item{config}{配置}

\item{src_view_name}{视图}

\item{img_table_name}{镜像文件}

\item{target_view_name}{目标视图}

\item{key_col}{关键字段}
}
\value{
返回值
}
\description{
测试删除数据
}
\examples{
edge_diff_del()
}
