% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_diff_pushToLog}
\alias{edge_diff_pushToLog}
\title{写入日志文件}
\usage{
edge_diff_pushToLog(
  config = cfg_edge,
  log_table_name = "t_icitem_view3_log",
  func_name = "func_icitem_max_ver",
  edge_diff_view_name = "t_icitem_view3_diff"
)
}
\arguments{
\item{config}{配置}

\item{log_table_name}{日志表}

\item{func_name}{版本}

\item{edge_diff_view_name}{}
}
\value{

}
\description{
写入日志文件
}
\examples{
edge_diff_pushLog()
}
