% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_img_pushToDel}
\alias{edge_img_pushToDel}
\title{对静像文件进行备份}
\usage{
edge_img_pushToDel(
  config = cfg_edge,
  img_table_name,
  imgDel_table_name,
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{img_table_name}{镜像文件}

\item{imgDel_table_name}{备份文件}

\item{pwd}{控制码}
}
\value{

}
\description{
对静像文件进行备份
}
\examples{
edge_img_pushToDel()
}
