% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_Log_create}
\alias{edge_Log_create}
\title{创建表的日志}
\usage{
edge_Log_create(
  config = cfg_edge,
  src_table_name = "t_icitem_view3_diff",
  target_table_name = "t_icitem_view3_log",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{src_table_name}{源表}

\item{target_table_name}{日志表}

\item{field_selector}{字段选择器}

\item{pwd}{控制码}
}
\value{
返顺值
}
\description{
创建表的日志
}
\examples{
edge_Log_create()
}
