% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_log_maxVersion}
\alias{edge_log_maxVersion}
\title{获取日志的最大版本号}
\usage{
edge_log_maxVersion(
  config = cfg_edge,
  log_table_name = "t_icitem_view3_log",
  func_name = "func_icitem_max_ver"
)
}
\arguments{
\item{config}{配置}

\item{log_table_name}{日志名}

\item{func_name}{函数名}
}
\value{
返回值
}
\description{
获取日志的最大版本号
}
\examples{
edge_log_maxVersion()
}
