% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_log_pushToCloud_input}
\alias{edge_log_pushToCloud_input}
\title{边向云端写入数据}
\usage{
edge_log_pushToCloud_input(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  input_table_name = "rds_icitem_view3_input",
  edge_diff_view_name = "t_icitem_view3_diff"
)
}
\arguments{
\item{config_edge}{边连接}

\item{config_cloud}{云连接}

\item{input_table_name}{表名}

\item{edge_diff_view_name}{差异表}
}
\value{
返回值
}
\description{
边向云端写入数据
}
\examples{
edge_log_pushToCloud_input()
}
