% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_log_pushToCloud_log}
\alias{edge_log_pushToCloud_log}
\title{边向云端写入日志}
\usage{
edge_log_pushToCloud_log(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  cloud_log_table_name = "rds_icitem_view3_log",
  edge_log_table_name = "t_icitem_view3_log"
)
}
\arguments{
\item{config_edge}{边配置}

\item{config_cloud}{云配置}

\item{cloud_log_table_name}{日志文件}

\item{edge_log_table_name}{日志文件}
}
\value{
返回值
}
\description{
边向云端写入日志
}
\examples{
edge_log_pushToCloud_log()
}
