% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{edge_view_pushToImg}
\alias{edge_view_pushToImg}
\title{将视图数据推入静像文件}
\usage{
edge_view_pushToImg(
  config = cfg_edge,
  src_view_name = "t_icitem_view3",
  img_table_name = "t_icitem_view3_img",
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{src_view_name}{视图}

\item{img_table_name}{静像文件}

\item{pwd}{密码}
}
\value{

}
\description{
将视图数据推入静像文件
}
\examples{
edge_view_pushToImg()
}
