% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{func_drop}
\alias{func_drop}
\title{删除函数对象}
\usage{
func_drop(config = cfg_edge, func_name = "func_icitem_max_ver")
}
\arguments{
\item{config}{配置}

\item{func_name}{函数名}
}
\value{
无返回值
}
\description{
删除函数对象
}
\examples{
func_drop()
}
