% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{Item_getUnAllocateNumbers}
\alias{Item_getUnAllocateNumbers}
\title{返回未分配的物料数据}
\usage{
Item_getUnAllocateNumbers(
  conn = vmrdspkg::conn_vm_erp_test(),
  n = 10,
  FPropType = "外购"
)
}
\arguments{
\item{conn}{连接}

\item{n}{返回数量}

\item{FPropType}{物料属性}
}
\value{
返回值
}
\description{
返回未分配的物料数据
}
\examples{
Item_getUnAllocateNumbers()
}
