% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{local_icitem_max_version}
\alias{local_icitem_max_version}
\title{获取最大号}
\usage{
local_icitem_max_version(cfg = cfg_lc)
}
\arguments{
\item{cfg}{配置文件}
}
\value{
返回值
}
\description{
获取最大号
}
\examples{
local_max_version()
}
