% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_createTable_backup}
\alias{md_createTable_backup}
\title{创建主数据备用表}
\usage{
md_createTable_backup(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name = "t_Item_rdsBak"
)
}
\arguments{
\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
创建主数据备用表
}
\examples{
md_createTable_backup()
}
