% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_createTable_rdsInput}
\alias{md_createTable_rdsInput}
\title{创建分配后的表结构}
\usage{
md_createTable_rdsInput(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name = "t_item_rdsInput"
)
}
\arguments{
\item{conn}{连接}

\item{table_name}{已分配表结构名称}
}
\value{
返回值
}
\description{
创建分配后的表结构
}
\examples{
md_createTable_rdsInput()
}
