% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_createTable_room}
\alias{md_createTable_room}
\title{建议基础资料的分配规则表}
\usage{
md_createTable_room(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
建议基础资料的分配规则表
}
\examples{
md_createTable_room()
}
