% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_createTable_unAllocated}
\alias{md_createTable_unAllocated}
\title{创建基础资料主数据的副表，用于存储待分配的物料}
\usage{
md_createTable_unAllocated(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name = "t_Item_rds"
)
}
\arguments{
\item{conn}{连接}
}
\value{
返回值
}
\description{
创建基础资料主数据的副表，用于存储待分配的物料
}
\examples{
md_createTable_unAllocated
}
