% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_material.R
\name{md_material_autoRun}
\alias{md_material_autoRun}
\title{物料同步}
\usage{
md_material_autoRun(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  config_meta = md_material_meta()
)
}
\arguments{
\item{config_edge}{边配置}

\item{config_cloud}{云配置}

\item{config_meta}{物料元数据}
}
\value{
返回值
}
\description{
物料同步
}
\examples{
md_material_autoRun()
}
