% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_material.R
\name{md_material_init}
\alias{md_material_init}
\title{定义物料主数据初始化}
\usage{
md_material_init(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  config_meta = md_material_meta()
)
}
\arguments{
\item{config_edge}{边配置}

\item{config_cloud}{云配置}

\item{config_meta}{元灵气}
}
\value{
返回值
}
\description{
定义物料主数据初始化
}
\examples{
md_material_init()
}
