% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_pushItem2UnAllocated_wg}
\alias{md_pushItem2UnAllocated_wg}
\title{将外购的物料推入待分配表}
\usage{
md_pushItem2UnAllocated_wg(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name_rds = "t_Item_rds",
  table_name_room = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{连接}

\item{table_name}{基础资料未分配表名}
}
\value{
返回值
}
\description{
将外购的物料推入待分配表
}
\examples{
 md_pushItem2UnAllocated_wg()
}
