% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_item.R
\name{md_pushItem2UnAllocated_zz}
\alias{md_pushItem2UnAllocated_zz}
\title{针对自制物料进行初始化隐藏处理}
\usage{
md_pushItem2UnAllocated_zz(
  conn = vmrdspkg::conn_vm_erp_test(),
  table_name_rds = "t_Item_rds",
  table_name_room = "t_item_rdsRoom"
)
}
\arguments{
\item{conn}{ERP链接信息}

\item{table_name_rds}{物料明细表}

\item{table_name_room}{待分配统计表}
}
\value{
返回值
}
\description{
针对自制物料进行初始化隐藏处理
}
\examples{
md_pushItem2UnAllocated_zz()
}
