% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mdm_autoRun}
\alias{mdm_autoRun}
\title{主数据同步}
\usage{
mdm_autoRun(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  config_meta = mdm_meta()
)
}
\arguments{
\item{config_edge}{边配置}

\item{config_cloud}{云配置}

\item{config_meta}{元数据}
}
\value{
返回值
}
\description{
主数据同步
}
\examples{
mdm_autoRun()
}
