% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_number.R
\name{mdm_getParentNumber}
\alias{mdm_getParentNumber}
\title{基础资料获取上级获取}
\usage{
mdm_getParentNumber(Numbers)
}
\arguments{
\item{Numbers}{明细编码使用.分隔}
}
\value{
返回值
}
\description{
基础资料获取上级获取
}
\examples{
mdm_getParentNumber()
mdm_getParentNumber('rds.01.001')
mdm_getParentNumber(c('rds.01.001.00001','rds.01.01'))
}
