% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mdm_init}
\alias{mdm_init}
\title{主数据创建服务}
\usage{
mdm_init(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  config_meta = mdm_meta()
)
}
\arguments{
\item{config_edge}{边配置}

\item{config_cloud}{云配置}

\item{config_meta}{元数据}
}
\value{
返回值
}
\description{
主数据创建服务
}
\examples{
mdm_init()
}
