% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_createFromCopy}
\alias{table_createFromCopy}
\title{复制表结构的数据}
\usage{
table_createFromCopy(
  config = cfg_edge,
  src_table_name = "t_icitem",
  target_table_name = "t_icitem_demo",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config}{配置}

\item{src_table_name}{数据源表名}

\item{target_table_name}{目的表名}

\item{field_selector}{字段选择器}

\item{pwd}{密码}
}
\value{
返回值
}
\description{
复制表结构的数据
}
\examples{
table_createFromCopy()
}
