% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_createFromCopy_edgeToCloud}
\alias{table_createFromCopy_edgeToCloud}
\title{将边的数据结构复制到云表上}
\usage{
table_createFromCopy_edgeToCloud(
  config_edge = cfg_edge,
  config_cloud = cfg_cloud,
  src_edge_tableName = "t_icitem_view3_log",
  target_cloud_tableName = "rds_icitem_view3_log",
  field_selector = NA,
  pwd = ""
)
}
\arguments{
\item{config_edge}{终的数据库配置}

\item{config_cloud}{云的数据库配置}

\item{src_edge_tableName}{复制来源，边的表名}

\item{target_cloud_tableName}{目标数据表，云的数据}

\item{field_selector}{字段选择器}

\item{pwd}{控制码}
}
\value{
返回值
}
\description{
将边的数据结构复制到云表上
}
\examples{
table_createFromCopy_edgeToCloud()
}
