% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_drop}
\alias{table_drop}
\title{删除表结构}
\usage{
table_drop(config = cfg_edge, table_name = "t_icitem_view", pwd = "")
}
\arguments{
\item{config}{配置文件}

\item{table_name}{表名}
}
\value{
无返回值
}
\description{
删除表结构
}
\examples{
table_drop()
}
