% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_fieldList}
\alias{table_fieldList}
\title{字段列表选择器}
\usage{
table_fieldList(config = cfg_edge, table_name = "t_icitem", exclude = "")
}
\arguments{
\item{config}{配置}

\item{table_name}{表名}

\item{exclude}{不需要字段名}
}
\value{
返回值
}
\description{
字段列表选择器
}
\examples{
table_fieldList()
}
