% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_maxId}
\alias{table_maxId}
\title{获取最大号}
\usage{
table_maxId(
  config = cfg_edge,
  table_name = "t_icitem_log",
  field_name = "FVersion"
)
}
\arguments{
\item{cfg}{配置文件}
}
\value{
返回值
}
\description{
获取最大号
}
\examples{
table_maxId()
}
