% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_select}
\alias{table_select}
\title{进行表的选择}
\usage{
table_select(config = cfg_edge, table_name = "t_icitem_view")
}
\arguments{
\item{config}{配置}

\item{table_name}{表或视图名}
}
\value{
返回值
}
\description{
进行表的选择
}
\examples{
table_select()
}
