% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_selectBy_Status}
\alias{table_selectBy_Status}
\title{按状态查询表}
\usage{
table_selectBy_Status(
  config = cfg_edge,
  table_name = "t_icitem_view",
  FStatus = "FIsDo",
  FValue = 0
)
}
\arguments{
\item{config}{配置文件}

\item{table_name}{表名}

\item{FStatus}{状态字段名}

\item{FValue}{状态值}
}
\value{
返回值
}
\description{
按状态查询表
}
\examples{
table_selectBy_Status()
}
