% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_setBy_Status}
\alias{table_setBy_Status}
\title{更新表的状态整数型}
\usage{
table_setBy_Status(
  config = cfg_edge,
  table_name = "t_icitem_view",
  FStatus = "FIsDo",
  FFromValue = 0,
  FToValue = 1
)
}
\arguments{
\item{config}{配置文件}

\item{table_name}{表名}

\item{FStatus}{状态字段}

\item{FFromValue}{原值}

\item{FToValue}{更改值}
}
\value{
返回值
}
\description{
更新表的状态整数型
}
\examples{
table_setBy_Status()
}
