% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{table_truncate}
\alias{table_truncate}
\title{针对表进行清空拉个口子要见血}
\usage{
table_truncate(config = cfg_edge, table_name = "t_dataTest", pwd = "")
}
\arguments{
\item{config}{配置}

\item{table_name}{表名}

\item{pwd}{密码}
}
\value{
无返回值
}
\description{
针对表进行清空拉个口子要见血
}
\examples{
table_truncate()
}
