% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tableInfo_queryField}
\alias{tableInfo_queryField}
\title{查询表的元数据数据}
\usage{
tableInfo_queryField(
  config = cfg_edge,
  table_name = "t_icitem",
  field_selector = NA
)
}
\arguments{
\item{config}{配置}

\item{table_name}{表名}
}
\value{
字段列表
}
\description{
查询表的元数据数据
}
\examples{
tableInfo_queryField()
}
