% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{view_create}
\alias{view_create}
\title{视图创建}
\usage{
view_create(
  config = cfg_edge,
  table_name = "t_icitem",
  view_suffix = "_view",
  selector = c("FItemID", "FName")
)
}
\arguments{
\item{config}{配置文件}

\item{table_name}{来源表名}

\item{view_suffix}{视图后缀}

\item{selector}{字段列表}
}
\value{
无返回值
}
\description{
视图创建
}
\examples{
view_create()
}
