% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{view_drop}
\alias{view_drop}
\title{删除视图}
\usage{
view_drop(config = cfg_edge, view_name = "t_icitem_view")
}
\arguments{
\item{config}{配置}

\item{view_name}{视图名称}
}
\value{
无返回值
}
\description{
删除视图
}
\examples{
view_drop()
}
