% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{view_select}
\alias{view_select}
\title{进行表的选择}
\usage{
view_select(config = cfg_edge, view_name = "t_icitem_view")
}
\arguments{
\item{config}{配置}

\item{view_name}{视图名}
}
\value{
返回值
}
\description{
进行表的选择
}
\examples{
table_select()
}
