% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{view_tableInfo_init}
\alias{view_tableInfo_init}
\title{针对表数据进行视频化}
\usage{
view_tableInfo_init(config = cfg_edge)
}
\arguments{
\item{config}{配置文件}
}
\value{
返回值
}
\description{
针对表数据进行视频化
}
\examples{
view_tableInfo_init()
}
